feat(refs: DPLAN-17697): move findInactivityDeletionCandidates to UserRepository
