+++ demosplan/DemosPlanCoreBundle/Repository/UserRepository.php
@@ -11,11 +11,13 @@
 namespace demosplan\DemosPlanCoreBundle\Repository;
 
 use Closure;
+use DateTimeInterface;
 use DemosEurope\DemosplanAddon\Contracts\Entities\CustomerInterface;
 use DemosEurope\DemosplanAddon\Contracts\Entities\OrgaInterface;
 use DemosEurope\DemosplanAddon\Contracts\Repositories\UserRepositoryInterface;
 use demosplan\DemosPlanCoreBundle\Entity\CoreEntity;
 use demosplan\DemosPlanCoreBundle\Entity\User\Address;
+use demosplan\DemosPlanCoreBundle\Entity\User\AiApiUser;
 use demosplan\DemosPlanCoreBundle\Entity\User\Customer;
 use demosplan\DemosPlanCoreBundle\Entity\User\Department;
 use demosplan\DemosPlanCoreBundle\Entity\User\FunctionalUser;
@@ -756,5 +758,38 @@ class UserRepository extends CoreRepository implements ArrayInterface, ObjectInt
         }
 
         return $qb;
+    }
+
+    /**
+     * Returns active users whose effective inactivity reference (`lastLogin` if set,
+     * else `createdDate`) is at or before the cutoff. Excludes deleted users, the
+     * AI API user (by login — its row ID is random per project), and any further
+     * protected IDs supplied by the caller (typically the anonymous-user constant
+     * plus project-specific protected accounts).
+     *
+     * @param list<string> $protectedUserIds
+     *
+     * @return list<User>
+     */
+    public function findInactivityDeletionCandidates(
+        DateTimeInterface $cutoff,
+        array $protectedUserIds,
+    ): array {
+        $queryBuilder = $this->getEntityManager()->createQueryBuilder()
+            ->select('u')
+            ->from(User::class, 'u')
+            ->where('u.deleted = false')
+            ->andWhere('COALESCE(u.lastLogin, u.createdDate) <= :cutoff')
+            ->andWhere('u.login != :aiApiUserLogin')
+            ->setParameter('cutoff', $cutoff)
+            ->setParameter('aiApiUserLogin', AiApiUser::AI_API_USER_LOGIN);
+
+        if ([] !== $protectedUserIds) {
+            $queryBuilder
+                ->andWhere('u.id NOT IN (:protectedUserIds)')
+                ->setParameter('protectedUserIds', $protectedUserIds);
+        }
+
+        return $queryBuilder->getQuery()->getResult();
     }
 }
